requirements.txt
tensorflow==2.13.0
streamlit
pillow
requests
numpy
scikit-learn
google-generativeai
